fix(customquotas): simplify error logging
